질문 데이터를 완전히 삭제하는 것이 아니라 데이터의 상태를 삭제 상태(deleted - boolean type)로 변경한다. 로그인 사용자와 질문한 사람이 같은 경우 삭제 가능하다. 답변이 없는 경우 삭제가 가능하다. 질문자와 답변 글의 모든 답변자 같은 경우 삭제가 가능하다. 질문을 삭제할 때 답변 또한 삭제해야 하며, 답변의 삭제 또한 삭제 상태(deleted)를 변경한다. 질문자와 답변자가 다른 경우 답변을 삭제할 수 없다. 질문과 답변 삭제 이력에 대한 정보를 DeleteHistory를 활용해 남긴다.

+++ src/main/java/codesquad/service/QnaService.java
@@ -62,11 +62,11 @@ public class QnaService {
     public void deleteQuestion(User loginUser, long questionId) throws CannotDeleteException {
         Question originalQuestion = questionRepository.findById(questionId).orElseThrow(EntityNotFoundException::new);
 
-        if (!originalQuestion.isOwner(loginUser)) {
-            throw new UnAuthorizedException();
-        }
+//        questionRepository.delete(originalQuestion);
 
-        questionRepository.delete(originalQuestion);
+        List<DeleteHistory> toSaveHistories = originalQuestion.delete(loginUser);
+        deleteHistoryService.saveAll(toSaveHistories);
+
     }
 
     public Iterable<Question> findAll() {
@@ -86,6 +86,10 @@ public class QnaService {
     @Transactional
     public void deleteAnswer(User loginUser, long id) {
         Answer savedAnswer = answerRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        if (!savedAnswer.isOwner(loginUser)) {
+            throw new UnAuthorizedException("답변 작성자와 로그인 유저가 다름");
+        }
+
         answerRepository.delete(savedAnswer);
     }
